class name typos and code formatting
